fix: don't use vim.schedule to defer window update
Previously, `vim.schedule` was used to make sure that the window updates
run after all other autocommands had fired, which was a workaround for a
specific issue with NvimTree. Since `vim.schedule` is a hacky and
unpredictable way of fixing that issue and has been causing issues with
multiple other things, it's been removed in this commit.

Since this also means that the current window no longer needs to be
updated after the update triggers, `update_all_windows` has been changed
back to `update_inactive_windows`