refactor: remove component truncation
The truncation system still has a lot of flaws such as not working with
Vim statusline modifiers, and there is no straightforward way to fix it.
It also greatly increased the complexity of the codebase. It also caused
a major performance downgrade. All these things combined warranted the
removal of this experimental feature. There has been [an Issue made in
the Neovim GitHub repo](https://github.com/neovim/neovim/issues/15849)
which will greatly assist in implementing this feature. But until that's
implemented in Neovim, this feature will be removed.

BREAKING CHANGE: Smart component truncation has been removed and all
configuration options that make use of it no longer work.