feat: make components table linear
Instead of having the components table contain a list of sections which in turn contain a list of components, make the components table contain a list of components directly and turn the section separator into a provider named `spacer`. This greatly enhances the configuration capabilities of Feline by making section gaps and their highlighting more explicit, while also allowing for section gaps to use the `default_hl` value by default. It also greatly simplifies the truncation code.

BREAKING CHANGE: The components table is now configured differently. To switch to the new configuration format, just combine all your sections into a single table and put the `spacer` provider in where your section gaps are supposed to be.