refactor: remove `default_hl`
The `default_hl` option is mostly unnecessary since StatusLine and StatusLineNC can easily be configured manually by the user. And since the default value of `default_hl` was using the same colors, it made the active statusline have '^' characters in it, which wasn't intended behavior for the users. In order to solve that, the `default_hl` option was removed.

BREAKING CHANGE: `default_hl` is now removed. Set the `StatusLine` and `StatusLineNC` highlights manually to set the highlights for the active and inactive statusline now.