fix: allow reloading feline configuration
Previously, Feline configuration couldn't be reloaded due to the custom providers from the previous configuration being cached, causing an error if a provider with the same name was in the new configuration as well. This commit fixes that by unloading the feline providers Lua module before loading it, to prevent its values from being cached.

Closes #229